Move choice extension functions outside of interactor so they can be reused
